+++ Common/src/main/java/fuzs/diagonalblocks/common/impl/config/CommonConfig.java
@@ -1,0 +1,10 @@
+package fuzs.diagonalblocks.common.impl.config;
+
+import fuzs.puzzleslib.common.api.config.v3.Config;
+import fuzs.puzzleslib.common.api.config.v3.ConfigCore;
+
+public class CommonConfig implements ConfigCore {
+    @Config(description = "When placing blocks use the diagonal variants instead of the non-diagonal blocks.",
+            gameRestart = true)
+    public boolean diagonalBlocksByDefault = true;
+}