+++ Common/src/main/java/fuzs/diagonalblocks/common/impl/data/tags/package-info.java
@@ -1,4 +1,4 @@
 @ApiStatus.Internal
-package fuzs.diagonalblocks.common.impl.data;
+package fuzs.diagonalblocks.common.impl.data.tags;
 
 import org.jetbrains.annotations.ApiStatus;