+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v26.1.0-mc26.1.x] - 2026-04-30
 
+### Added
+
+- Add a config setting for placing non-diagonal blocks by default
+- Allow for toggling the swapping between diagonal and non-diagonal blocks mechanic in the config
+
 ### Changed
 
 - Update to Minecraft 26.1.x